modifique aplicacion Vendedor, Envios, bd y gateway con las nuevas apis

+++ API-Inventario/src/main/java/com/example/API_Inventario/repository/InventarioRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface InventarioRepository extends JpaRepository<Inventario, Long> {
+public interface InventarioRepository extends JpaRepository<Inventario, Integer> {
    
 }
